refactor: split LiveTutor.tsx into co-located sub-components (Related to #962)
Move 1101-line monolith into live-tutor/ sub-directory:
- LiveTutor.tsx — thin orchestrator (state + hooks + logic, 723 lines)
- ParagraphCard.tsx — per-paragraph card with controls + summary (316 lines)
- TutorFeedbackPanel.tsx — right panel with transcript + diff display (112 lines)
- liveTutorRenderHelpers.tsx — pure helpers (calcSpeakingProgress, renderLineWithDiff)
- liveTutorTypes.ts — shared LineResult + ParagraphSummaryData types
- index.tsx — re-export for backwards-compatible imports

Original reading-steps/LiveTutor.tsx reduced to a 3-line re-export stub.
No behavior changes — pure structural refactor.
