+++ frontend/src/components/reading-steps/live-tutor/index.tsx
@@ -1,0 +1,4 @@
+// Re-export LiveTutor as default for backwards-compatible imports
+// e.g. import LiveTutor from '.../reading-steps/LiveTutor'
+// still works via the stub in the parent directory.
+export { default } from './LiveTutor';
